requirements.txt
anyio==3.6.1
APScheduler==3.9.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
decorator==4.4.2
ffmpeg-python==0.2.0
future==0.18.2
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
imageio==2.21.3
imageio-ffmpeg==0.4.7
isort==5.10.1
moviepy==1.0.3
numpy==1.23.3
Pillow==9.2.0
proglog==0.1.10
python-dotenv==0.21.0
python-telegram-bot==20.0a4
pytube==12.1.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
tqdm==4.64.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
